requirements.txt
streamlit>=1.32.0
google-auth>=2.28.0
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.118.0
PyPDF2>=3.0.1
python-docx>=1.1.0
requests>=2.31.0
pandas>=2.2.0
numpy>=1.26.0
networkx>=3.2.0
plotly>=5.18.0
kaleido>=0.2.1
supabase>=2.3.0
jinja2>=3.1.3
python-dotenv>=1.0.1
tenacity>=8.2.3
